Начало добавления проекта для проверки модели

+++ GraduationProject/ModelObjects/IObjects/IFillet.cs
@@ -3,5 +3,5 @@
 public interface IFillet
 {
     public string Name { get; set; }
-    public double Radius { get; set; }
+    public decimal Radius { get; set; }
 }
