+++ GraduationProject/ModelObjects/IObjects/ISketchObjects/IPoints/IPoint.cs
@@ -2,11 +2,32 @@
 
 public interface IPoint
 {
+    /// <summary>
+    /// Координаты для вывода
+    /// </summary>
     public string Coordinate { get; set; }
-    public double XStart { get; set; }
-    public double YStart { get; set; }
-    public double ZStart { get; set; }
-    public double XEnd { get; set; }
-    public double YEnd { get; set; }
-    public double ZEnd { get; set; }
+    /// <summary>
+    /// Начало координат по оси X
+    /// </summary>
+    public decimal XStart { get; set; }
+    /// <summary>
+    /// Начало координат по оси Y
+    /// </summary>
+    public decimal YStart { get; set; }
+    /// <summary>
+    /// Начало координат по оси Z
+    /// </summary>
+    public decimal ZStart { get; set; }
+    /// <summary>
+    /// Конец координат по оси X
+    /// </summary>
+    public decimal XEnd { get; set; }
+    /// <summary>
+    /// Конец координат по оси Y
+    /// </summary>
+    public decimal YEnd { get; set; }
+    /// <summary>
+    /// Конец координат по оси Z
+    /// </summary>
+    public decimal ZEnd { get; set; }
 }
