+++ GraduationProject/ModelObjects/IObjects/ISketchObjects/IUserPoint.cs
@@ -3,7 +3,7 @@
 public interface IUserPoint
 {
     public string Coordinate { get; set; }
-    public double X { get; set; }
-    public double Y { get; set; }
-    public double Z { get; set; }
+    public decimal X { get; set; }
+    public decimal Y { get; set; }
+    public decimal Z { get; set; }
 }
